Server: add support for asymmetric signatures
This is controlled by the signature_asymmetric_keys configuration value.
If this value is set to private keys, then the signature scheme changes
from symmetric to asymmetric using the set keys.

The scheme used is ed25519(sha512(m)).

People can generate keypairs for this by running:
svix-server asymmetric-key generate

Please note, that asymmetric signatures are a few orders of magnitude
slower than symmetric signatures, and also more complex. So only use it
if it makes sense for your environment.

Fixes #501