Server: make asymmetric key support dynamic (instead of cfg)
In the previous commit we added support for signing webhooks using
asymmetric signatures. Though the way we did it, the keys were globally
defined in the config. While simple and easy, it's not a good solution
because of security.

If you use the same key globally, then attackers could just create their
own endpoints and set the URLs to whatever they want, and the Svix
server will blindly sign the request. This is a problem whether we also
sign the URL or not. The only way to slightly alleviate it is by signing
the endpoint ID, but that becomes a pain to handle and for no good
reason.

This change therefore makes the asymmetric key dynamic, and they can be
set per endpoint, and each endpoint can even switch (by rotating)
between one method to the other. All live.

We achieve this by adding support for a new prefix for endpoint secrets:
`whsk_` which indicates an asymmetric secret key. When you then get the
secret it returns the public key part (which is prefixed with `whpk_`).

We have also added a configuration variable that lets you set the
default key type being set, when a specific key isn't passed.